add: provide storage access to handler functions
Improve storage functionality and give access to handler functions.

Before, each of the functions that responsible to handle storage operations,
like insertion and read of the hosts and its details from the database, had
to connect first, and then do their operations. Every time.

This is an unefficient way to handle database connections.
Instead, providing a single, "centralized" way of handling the connection to
the database (like a sort of database layer), and then providing access
to the same database pool to the handlers would have positive effects on
speed and code readability and maintenance.